feat: move auth middleware logic into its own file
